# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TeamCloud",
    "description": "API for working with a TeamCloud instance.",
    "contact": {
      "name": "TeamCloud Dev Team",
      "url": "https://github.com/microsoft/TeamCloud/issues/new",
      "email": "Markus.Heiliger@microsoft.com"
    },
    "license": {
      "name": "TeamCloud is licensed under the MIT License",
      "url": "https://github.com/microsoft/TeamCloud/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/orgs/{organizationId}/scopes": {
      "get": {
        "tags": [
          "DeploymentScopes"
        ],
        "summary": "Gets all Deployment Scopes.",
        "operationId": "GetDeploymentScopes",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns all Deployment Scopes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentScopeListDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "DeploymentScopes"
        ],
        "summary": "Creates a new Deployment Scope.",
        "operationId": "CreateDeploymentScope",
        "parameters": [
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentScopeDefinition"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The new Deployment Scope was created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentScopeDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "409": {
            "description": "A Deployment Scope already exists with the ID provided in the request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/orgs/{organizationId}/scopes/{id}": {
      "get": {
        "tags": [
          "DeploymentScopes"
        ],
        "summary": "Gets a Deployment Scope.",
        "operationId": "GetDeploymentScope",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns a DeploymentScope.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentScopeDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A DeploymentScope with the id provided was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "DeploymentScopes"
        ],
        "summary": "Updates an existing Deployment Scope.",
        "operationId": "UpdateDeploymentScope",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentScope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The DeploymentScope was updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentScopeDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A Deployment Scope with the ID provided in the request body could not be found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "DeploymentScopes"
        ],
        "summary": "Deletes a Deployment Scope.",
        "operationId": "DeleteDeploymentScope",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The DeploymentScope was deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentScopeDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A DeploymentScope with the id provided was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/orgs/{organizationId}/scopes/{id}/authorize": {
      "put": {
        "tags": [
          "DeploymentScopes"
        ],
        "summary": "Authorize an existing Deployment Scope.",
        "operationId": "AuthorizeDeploymentScope",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentScope"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentScope"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentScope"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DeploymentScope"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The DeploymentScope was updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentScopeDataResult"
                }
              }
            }
          },
          "400": {
            "description": "A validation error occured.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "404": {
            "description": "A DeploymentScope with the id provided was not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResult"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ResultError": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "Unknown",
              "Failed",
              "Conflict",
              "NotFound",
              "ServerError",
              "ValidationError",
              "Unauthorized",
              "Forbidden"
            ],
            "type": "string"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ErrorResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResultError"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DeploymentScope": {
        "required": [
          "displayName",
          "id",
          "isDefault",
          "organization",
          "slug"
        ],
        "type": "object",
        "properties": {
          "organization": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "adapter": {
            "enum": [
              "AzureResourceManager",
              "AzureDevOps",
              "GitHub"
            ],
            "type": "string"
          },
          "managementGroupId": {
            "type": "string",
            "nullable": true
          },
          "subscriptionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "authorizeUrl": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "DeploymentScopeListDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeploymentScope"
            },
            "nullable": true,
            "readOnly": true
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeploymentScopeDefinition": {
        "required": [
          "displayName"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "isDefault": {
            "type": "boolean"
          },
          "managementGroupId": {
            "type": "string",
            "nullable": true
          },
          "subscriptionIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DeploymentScopeDataResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/DeploymentScope"
          },
          "location": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://login.microsoftonline.com//oauth2/v2.0/authorize",
            "tokenUrl": "https://login.microsoftonline.com//oauth2/v2.0/token",
            "scopes": {
              "openid": "Sign you in",
              "http://TeamCloud.aztcclitestsix/user_impersonation": "Access the TeamCloud API"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "oauth2": [
        "openid",
        "http://TeamCloud.aztcclitestsix/user_impersonation"
      ]
    }
  ]
}